pyproject.toml
[tool.poetry]
name = "minilab3synth"
version = "0.1.0"
description = ""
authors = ["Julien Desfossez <j@jdfz.org>"]
license = "GPLv2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.1"
pyaudio = "^0.2.14"
pyfluidsynth = "^1.3.3"
python-rtmidi = "^1.5.8"
ruff = "^0.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
